websocket & RPCHandler generic
