+++ xelis_common/src/prompt/mod.rs
@@ -237,9 +237,6 @@ impl Prompt {
         base = base.level(level.into());
 
         base.level_for("sled", log::LevelFilter::Warn)
-        .level_for("actix_server", log::LevelFilter::Warn)
-        .level_for("actix_web", log::LevelFilter::Warn)
-        .level_for("actix_http", log::LevelFilter::Warn)
         .level_for("mio", log::LevelFilter::Warn)
         .apply()?;
