+++ xelis_common/src/rpc_server/rpc_handler.rs
